requirements.txt
python-decouple
requests
#requests_cache
#requests[socks]

# validation of yaml/json data
cerberus

# webserver stuff
flask
waitress

# text to speech
faster_whisper
wyoming

trafilatura
youtube-transcript-api
# tts stuff
coqui-tts
spacy

#deepspeed

# scheduling recurring tasks
schedule

websockets

# wppconnect events api (for whatsapp)
python-socketio
python-socketio[client]

# get language from text
langid

bs4
yfinance
pyTelegramBotAPI
Flask

# image processing
Pillow